clean & comment: renamed radio/types -> radio/model.rs, dsp/demod.rs -> dsp/modes.rs to avoid confusion
