+++ rigflow-core/src/dsp/modes.rs
